Attempted SKIP_SOAK macro
The idea is to allow skipping of the soak time (e.g. for a faster print maybe?).

This idea did not work. I guess the macros are executed one after another. At the very least, this does allow for skipping the soak period prior to running PRINT_START (but not while it's running)